Create snowSkiing.c
Use iteration to find the longest path from the top of the hill to the foot of the hill. The hill is represented by a metric, and the foot is represented by the smallest number in the metric and the top is represented by the biggest number in the metric.